Add matmaul alias
